+++ Source/OxyPlot/Rendering/ScreenVector.cs
@@ -15,7 +15,7 @@ namespace OxyPlot
     /// <summary>
     /// Represents a vector defined in screen space.
     /// </summary>
-    public struct ScreenVector
+    public struct ScreenVector : IEquatable<ScreenVector>
     {
         /// <summary>
         /// The x-coordinate.
@@ -151,6 +151,16 @@ namespace OxyPlot
         public override string ToString()
         {
             return this.x + " " + this.y;
+        }
+
+        /// <summary>
+        /// Determines whether this instance and another specified <see cref="T:ScreenVector" /> object have the same value.
+        /// </summary>
+        /// <param name="other">The point to compare to this instance.</param>
+        /// <returns><c>true</c> if the value of the <paramref name="other" /> parameter is the same as the value of this instance; otherwise, <c>false</c>.</returns>
+        public bool Equals(ScreenVector other)
+        {
+            return this.x.Equals(other.x) && this.y.Equals(other.y);
         }
     }
 }